src/main/java/com/beetle/recipe/converters/NoteToNoteCommand.java
package com.beetle.recipe.converters;

import com.beetle.recipe.commands.NoteCommand;
import com.beetle.recipe.model.entity.Note;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * Created by jt on 6/21/17.
 */
@Component
public class NoteToNoteCommand implements Converter<Note, NoteCommand> {

    @Synchronized
    @Nullable
    @Override
    public NoteCommand convert(Note source) {
        if (source == null) {
            return null;
        }

        final NoteCommand noteCommand = new NoteCommand();
        noteCommand.setId(source.getId());
        noteCommand.setRecipeNotes(source.getNote());

        return noteCommand;
    }

}
